修复查找菜式成分不存在是处理错误

+++ src/main/java/com/bkit/fatdown/mappers/TbFoodElementRelationMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbFoodElementRelationMapper {
-    long countByExample(TbFoodElementRelationExample example);
+    Integer countByExample(TbFoodElementRelationExample example);
 
     int deleteByExample(TbFoodElementRelationExample example);
